Spring: Implement authentication against the user
Spring Boot adds per default a HTTP Basic authentication. This commit
changes the authentication to be form based. The authentication is done
against the signed up user. The user's password is stored with a
secure random salt and hashed with BCrypt.

I've decided to promote annotation based access control. I.e.
per default everything is accessible. Add @PreAuthorize to e.g. a
controller to control authorization. This commit includes as an
authorization example the account controller. The account controller
allows changing the password of the currently logged in user.

komarserjio/notejam#52