Spring: Fix error after creating a pad
Creating a pad did redirect to /pads/{id}?created. The created request
parameter is a hint for the controller to show a success message.
Unfortunately created collides with the identically named pad property
created. In combination that URI leads to an entity with an empty
created property. That triggers hibernates validation to raise an
exception.

See also: komarserjio/notejam#52